Files move, file rename, new tests, etc
